Update GraphQL endpoint to only accept one sample per well
For Ont samples
