+++ spec/requests/v2/plates_spec.rb
@@ -65,19 +65,8 @@ RSpec.describe 'GraphQL', type: :request do
             arguments: {
               barcode: "PLATE-1234"
               wells: [
-                {
-                  position: "A1"
-                  samples: [
-                    { name: "Sample 1 for A1" externalId: "ExtIdA1-1" }
-                    { name: "Sample 2 for A1" externalId: "ExtIdA1-2" tagGroupId: "Tag123" }
-                  ]
-                }
-                {
-                  position: "E7"
-                  samples: [
-                    { name: "Sample for E7" externalId: "ExtIdE7" }
-                  ]
-                }
+                { position: "A1" sample: { name: "Sample for A1" externalId: "ExtIdA1-1" } }
+                { position: "E7" sample: { name: "Sample for E7" externalId: "ExtIdE7" } }
               ]
             }
           }
@@ -92,8 +81,8 @@ RSpec.describe 'GraphQL', type: :request do
 
     it 'creates a plate with provided parameters' do
       plate = create(:plate_with_ont_samples, barcode: 'PLATE-1234', wells: [
-        { position: 'A1', samples: [ { name: 'Sample 1 for A1', external_id: 'ExtIdA1-1' }, { name: 'Sample 2 for A1', external_id: 'ExtIdA1-2' } ] },
-        { position: 'E7', samples: [ { name: 'Sample for E7', external_id: 'ExtIdE7' } ] }])
+        { position: 'A1', samples: [ { name: 'Sample for A1', external_id: 'ExtIdA1-1' } ] },
+        { position: 'E7', samples: [ { name: 'Sample for E7', external_id: 'ExtIdE7' } ] } ] )
 
       allow_any_instance_of(Ont::PlateFactory).to receive(:save).and_return(true)
       allow_any_instance_of(Ont::PlateFactory).to receive(:plate).and_return(plate)
@@ -107,8 +96,7 @@ RSpec.describe 'GraphQL', type: :request do
       expect(plate_json['barcode']).to eq('PLATE-1234')
 
       expect(plate_json['wells'][0]['materials']).to contain_exactly(
-        { 'sample' => { 'name' => 'Sample 1 for A1', 'externalId' => 'ExtIdA1-1' } },
-        { 'sample' => { 'name' => 'Sample 2 for A1', 'externalId' => 'ExtIdA1-2' } }
+        { 'sample' => { 'name' => 'Sample for A1', 'externalId' => 'ExtIdA1-1' } }
       )
       expect(plate_json['wells'][1]['materials']).to contain_exactly(
         { 'sample' => { 'name' => 'Sample for E7', 'externalId' => 'ExtIdE7' } }
